requirements.txt
edge-tts>=7.2.0
Pillow>=10.0.0